Create gsm_receive_rtl_mod_server.py
Code is work in progress and might not work on your modem or phone, Ive only tested this with my modem and phone but should work with others with little modification.

I havent looked at this code in awhile and if something doesnt seem right let me know and I will try fix it buttons might be missed placed or used for other functions. 
Links tmsi to phone number and sends silent sms etc.

needs to be placed in airprobe/gsm-receiver/src/python folder along with functions.py and sms_functions.py files will also need to be chmod. check out videos on youtube of code in action. search for kali gsm.