Create gsmstack_server.c
you will need to version of gsmstack if you want to use the gsm_receive_rtl_server mod to work.

rename gsmstack.c to gsmstack_bak.c and rename gsmstack_server.c to gsmstack.c and make gsm-receiver again.

if you want to go back to previous gsmtack just rename gsmstack_bak.c to gsmstack.c and make gsm-receiver again.

Server ip and port is hardcoded here so change to the server socket ip/port that server is listening on.

this will send all the data from output to any server that its connected to. so you can send data to an android phone and with a basic server app.  This would also come in useful for parsing the gsm data to extract any data thats in the packets in realtime with a server socket app. for example if you send this data to a py program with a server socket you could filter the data to target a particular tmsi in realtime to see if it gets paged.